HW#8 second task с разными файлами-классами
