+++ src/Planets/Earth.java
@@ -1,0 +1,18 @@
+package Planets;
+
+public class Earth extends MainPlanets.Planet implements MainPlanets.PlanetInterface {
+    int number = 3;
+    double mas = 597360000;
+    float G = (float) 9.780327;
+
+    int getNumber() {
+        return number;
+    }
+    double getMas() {
+        return mas;
+    }
+
+    public float getG() {
+        return G;
+    }
+}
