+++ src/Planets/Mercury.java
@@ -1,0 +1,18 @@
+package Planets;
+
+public    class Mercury extends MainPlanets.Planet implements MainPlanets.PlanetInterface {
+    int number = 1;
+    double mas = 97360000;
+    float G = (float) 3.7;
+
+    int getNumber() {
+        return number;
+    }
+    double getMas() {
+        return mas;
+    }
+
+    public float getG() {
+        return G;
+    }
+}
