+++ src/Planets/PlanetInterface.java
@@ -1,0 +1,5 @@
+package Planets;
+
+public interface PlanetInterface {
+    float getG(); //как-бы аьстрактный метод получения G
+}
