+++ src/Planets/Uranus.java
@@ -1,0 +1,18 @@
+package Planets;
+
+public class Uranus extends MainPlanets.Planet implements MainPlanets.PlanetInterface {
+        int number = 7;
+        double mas = 41260000;
+        float G = (float) 8.87;
+
+        int getNumber() {
+            return number;
+        }
+        double getMas() {
+            return mas;
+        }
+
+        public float getG() {
+            return G;
+        }
+    }
